backend/requirements.txt
# LawyerClaw Backend Dependencies
# Based on conda "justice" environment (Python 3.12)
# Pinned to exact versions that are known to work

# Web Framework
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Werkzeug==3.1.8

# Database
SQLAlchemy==2.0.36

# LLM Provider SDK
openai==2.31.0
anthropic==0.55.0
dashscope==1.23.6
zhipuai==2.1.5.20250106

# HTTP / Network
httpx==0.27.0
httpx-sse==0.4.1
requests==2.33.0
websocket-client==1.8.0
sse-starlette==3.3.4

# Async / Concurrency
nest-asyncio==1.6.0
tenacity==9.0.0

# Data Validation & Config
pydantic==2.9.0
pydantic-settings==2.13.1
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.2

# Image Processing
Pillow==11.0.0

# Document Parsing
PyPDF2==3.0.1
pdfplumber==0.11.9
python-docx==1.1.0
openpyxl==3.1.2
PyMuPDF==1.23.26
Markdown==3.5

# Encryption
pycryptodome==3.23.0

# RAG / Vector Database
chromadb==1.5.5

# Search
tavily-python==0.5.0

# System Tray
pystray==0.19.5
six==1.16.0

# Utilities
tqdm==4.66.5

# ASGI (optional, for async endpoints)
uvicorn==0.32.0
starlette==1.0.0
